requirements .txt
requests==2.28.0
prettytable==3.3.0
dnspython==2.2.1